requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
marshmallow==3.20.1
email-validator==2.1.0
python-dotenv==1.0.0
pytest==7.4.2
Werkzeug==2.3.7 